chore: moved seed_programs to its own command
