+++ README.md
@@ -28,6 +28,7 @@ Note: run all this in the repo's root directory
         python ./manage.py migrate
         python ./manage.py loaddata cpho/fixtures/dimension_lookups.yaml
         python ./manage.py loaddata cpho/fixtures/periods.yaml
+        python -m manage seed_programs --mode reset
         python ./manage.py runscript cpho.scripts.dev
         ```
 10. `python manage.py runserver`
